Remove xdo_t members that were removed upstream
At some point the upstream xdotool removed the modmap and keymap members
from the xdo struct, which breaks the ABI. The Debian maintainer wrote a
patch to keep it in and preserve ABI compatibility.

Since we are only dealing with the a pointer to an xdo_t, it should be
OK to remove it and still have it work on Debian.